requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
reportlab==4.2.0
pandas==2.2.2
openpyxl==3.1.3
xlrd==2.0.1
gunicorn==21.2.0
Pillow==10.3.0
psycopg2-binary==2.9.9
